move MangedProps logic to a higher level untils file
